Create TictactoeGUI.py
Hello coders, and obviously hooomaaanssss, this is Wangyal! 
I was inactive since so so long and I am really sorry for that. So, I am finally back with a new program, as you can see in the thumbnail itself. Hope you like it, and don't forget to like, share, comment.